Make currency enums all caps
When reading from defaults.yml, #[strum(ascii_case_insensitive) does not
have the desired effect. The case of the currencies in the file must
exactly match the case of the enum names.

To prevent errors, the currencies in defaults.yml are written in ALL
CAPS when the file is being created, and the enums names are in ALL CAPS
as well.

This won't however prevent the user from changing the currencies in
default.yml to lowercase letters, which would then result in a parsing
error.